requirements.txt
pandas==2.0.3
psycopg2-binary==2.9.6
jupyter==1.0.0
matplotlib==3.7.1
seaborn==0.12.2
python-dotenv==1.0.0
streamlit==1.32.0
plotly==5.18.0
sqlalchemy==2.0.25
ipykernel==6.28.0